requirements.txt
SimpleITK==1.2.4
h5py==2.10.0
keras==2.2.4
pandas==0.24.2
scipy==1.2.1
numpy==1.16.4
scikit-image==0.16.2
tensorflow-gpu==1.13.1
